Refactor StyledTextEditor: Move bubble menu toolbar in separate component
