+++ README.md
@@ -20,7 +20,7 @@ cd ./dev/
 rush build    # Will build all the required packages.
 rush bundle   # Will prepare bundles.
 rush docker:build   # Will build Docker containers for all applications in the local Docker environment.
-docker-compose up -d --force-recreate # Will set up all the containers
+docker compose up -d --force-recreate # Will set up all the containers
 ```
 
 By default, Docker volumes named dev_db, dev_elastic, and dev_files will be created for the MongoDB, Elasticsearch, and MinIO instances.
