+++ packages/theme/styles/_text-editor.scss
@@ -6,5 +6,5 @@
 // need to override editor's bubble max-width
 // due to https://github.com/atomiks/tippyjs/issues/451
 .tippy-box {
-  max-width: 450px !important;
+  max-width: 38rem !important;
 }
